fix(core): address HIGH and IMPORTANT audit findings
H1: exec:// token source now checks UnsafeFeatures allowlist.
H2: SOCKS5 bind address extra byte calculation fixed (IPv4:0, IPv6:12).
H3: WebSocket accept-from-peeked now feeds pipelined bytes into WsByteStream.
H4: WsByteStream pong payload capped at 125 bytes with warning.
H5: Add golden-value tests for V2 handshake transcript hash.
H6: NOT APPLICABLE - delay_tx.send().expect() already removed from frp-core.
I1: into_split() debug_assert! replaced with assert!.
I2: TcpMuxConfig keepalive applied via timeout-based poll loops.
I3: AeadStreamReader::read_exact fixed to detect EOF and avoid infinite loop.
I4: Add CipherStream round-trip/empty/corrupt/large tests.
I5: Symmetric shutdown added to bridge_encrypted (matching bridge_plain).
I6: Admin auth middleware now returns 500 if extension injection fails.
I7: NOT APPLICABLE - spin::Mutex already removed from frp-core.
I8: Add zeroize_string() helper for token memory zeroization.
M1: Fix typo 'configures' -> 'configuration' in CLI help.
M2: V1 length field type made consistent (u64 both sides).
M3: tracing::error! -> warn! for admin auth config warning.
M4: Add Default derive for VisitorConfig.
M5: Extract shared base64_encode helper in transport.rs.
M10: Self-signed TLS cert changed from IsCa::Ca to NoCa.

+++ frp-core/src/protocol.rs
@@ -13,13 +13,13 @@ pub async fn write_v1_frame<W: AsyncWriteExt + Unpin>(
     let payload = serde_json::to_vec(msg)
         .map_err(|e| crate::Error::Protocol(format!("serialize V1 msg: {e}")))?;
 
-    if payload.len() as i64 > V1_MAX_MSG_LENGTH {
+    if payload.len() as u64 > V1_MAX_MSG_LENGTH as u64 {
         return Err(crate::Error::Protocol("V1 message too large".into()));
     }
 
     let mut buf = Vec::with_capacity(V1_HEADER_LEN + payload.len());
     buf.push(type_byte);
-   buf.extend_from_slice(&(payload.len() as i64).to_be_bytes());
+   buf.extend_from_slice(&(payload.len() as u64).to_be_bytes());
    buf.extend_from_slice(&payload);
 
     tracing::trace!(
